# packages/shared-types/src/types/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pinaka API",
    "version": "1.0.0",
    "description": "API documentation generated from canonical schema registry"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/v1/vendors": {
      "get": {
        "summary": "List Vendors",
        "tags": [
          "Vendor"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of Vendors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Vendor",
        "tags": [
          "Vendor"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vendor created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vendors/{id}": {
      "get": {
        "summary": "Get Vendor by ID",
        "tags": [
          "Vendor"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vendor details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Vendor",
        "tags": [
          "Vendor"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vendor updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Vendor",
        "tags": [
          "Vendor"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Vendor deleted"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VendorCreate": {
        "type": "object",
        "description": "Schema for creating a Vendor"
      },
      "VendorUpdate": {
        "type": "object",
        "description": "Schema for updating a Vendor"
      },
      "VendorResponse": {
        "type": "object",
        "description": "Response schema for Vendor"
      },
...
    }
  },
...
}